dividindo dependencias melhor entre componentes
